api/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
cryptography==42.0.5
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Pillow==10.2.0
qrcode==7.4.2
pyotp==2.9.0
python-dateutil==2.9.0
itsdangerous==2.1.2
pycryptodome==3.20.0